moved data to db.js and create ShowAllProjects component
